fix: R4MU Ausbaukosten - Bugfixes für Netzausbaukosten-Berechnung
Stand: 94/106 Grids erfolgreich berechenbar, 12 noch offen.

=== FUNKTIONIERT ===

1. reinforce_measures.py - Voltage-Reinforcement-Bug behoben:
   Wenn node_2_3 bereits der Representative war (direkt an Station),
   wurde nur das erste Leitungssegment verstärkt statt ALLER Segmente
   auf dem Pfad zum kritischen Knoten. Jetzt werden alle Segmente
   iteriert und verstärkt.

2. generators_import.py - Robusterer Generator-Import:
   - Strenge Validierung (Anzahl/Kapazität) von raise auf warning
     geändert für die obere oedb()-Funktion (Zeile 984-996)
   - Einzelne Generatoren/CHP-Anlagen werden bei Fehler übersprungen
     statt den gesamten Import abzubrechen (try/except pro Generator)

3. heat_pump_import.py - Robusterer Wärmepumpen-Import:
   - Schema-Fix: egon_etrago_bus/link von "supply" auf "grid"
   - Zentrale Wärmepumpen mit p_set<=0 werden übersprungen
   - ValueError bei zu hoher Leistung wird gefangen statt abzubrechen

4. config_grid_expansion_default.cfg - Leicht erhöhte Spannungsgrenzen:
   - mv_lv_max_v_rise/drop: 0.10 -> 0.11
   - lv_max_v_rise: 0.035 -> 0.04
   - lv_max_v_drop: 0.065 -> 0.07

5. timeseries.py - DataFrame dtype fix:
   - power_scaling DataFrame mit dtype=float initialisiert

6. topology.py - Numerischer Typ-Fix:
   - Neue Leitungen: pd.to_numeric auf alle Spalten angewendet
     (verhindert Object-dtype durch gemischte Typen)

=== NOCH OFFEN (12 Grids) ===

8x Voltage >10% (35538, 36379, 36431, 38609, 40059, 40089, 40253, 40351):
   -> Lösung: split_voltage_band=True im finalen Check (reinforce_grid.py:665)
      und/oder activate_cost_results_disturbing_mode=True

1x Generator-Count (36309):
   -> Wird durch den warning-Fix oben bereits gelöst

1x Fehlende PF-Results (36371):
   -> Lösung: analyze() vor enhanced_reinforce_grid() aufrufen

1x Kaputtes Grid (37580):
   -> Grid-Topologie fehlt komplett, muss neu generiert oder übersprungen werden

1x Duplicate Labels (40446):
   -> Lösung: Deduplizierung in generators_import.py vor gens_df.update()

+++ edisgo/network/timeseries.py
@@ -981,7 +981,7 @@ class TimeSeries:
         worst_case_scale_factors = configs["worst_case_scale_factor"]
         # get power scaling factors for different use cases, voltage levels and
         # feed-in/load case
-        power_scaling = pd.DataFrame(columns=sectors)
+        power_scaling = pd.DataFrame(columns=sectors, dtype=float)
         for s in sectors:
             for case in cases:
                 for voltage_level in ["mv", "lv"]:
